requirements.txt
#Heroku
Django>=1.5.1
gunicorn>=0.17.4
psycopg2==2.4.5
dj-database-url==0.2.0
wsgiref==0.1.2
gevent>=0.13.8
South>=0.7.4
celery>=2.3.3
django-celery>=3.0.11
django-picklefield

#Site Specific
Markdown>=2.0.3
newrelic

#Cache
django-redis-cache